nighty: Do not rebuild an existing nightly.

+++ nightly/build.sh
@@ -67,31 +67,35 @@ popd
 PY3_VERSION=$(./lib/py3-linux-x86_64/python -O distribute.py --print-version --nightly)
 PY2_VERSION=$(./lib/py2-linux-x86_64/python -O distribute.py --print-version --nightly)
 
-# Build the distribution.
-./lib/py3-linux-x86_64/python -O distribute.py --pygame $BASE/pygame_sdl2 $DISTRIBUTE_ARGS --nightly
-./lib/py2-linux-x86_64/python -O distribute.py --pygame $BASE/pygame_sdl2 $DISTRIBUTE_ARGS --nightly
+if [ -e renpy/dl/$PY3_VERSION/branch.txt -a -e renpy/dl/$PY2_VERSION/branch.txt ]; then
+    echo "Already built."
+else
+    # Build the distribution.
+    ./lib/py3-linux-x86_64/python -O distribute.py --pygame $BASE/pygame_sdl2 $DISTRIBUTE_ARGS --nightly
+    ./lib/py2-linux-x86_64/python -O distribute.py --pygame $BASE/pygame_sdl2 $DISTRIBUTE_ARGS --nightly
 
-# Copy VCS data to distribution directories.
-pushd $BASE
-cp tmp/vcs7.json renpy/dl/$PY2_VERSION/vcs.json
-cp tmp/vcs8.json renpy/dl/$PY3_VERSION/vcs.json
+    # Copy VCS data to distribution directories.
+    pushd $BASE
+    cp tmp/vcs8.json renpy/dl/$PY3_VERSION/vcs.json
+    cp tmp/vcs7.json renpy/dl/$PY2_VERSION/vcs.json
 
-echo $BRANCH > renpy/dl/$PY2_VERSION/branch.txt
-echo $BRANCH > renpy/dl/$PY3_VERSION/branch.txt
-popd
+    echo $BRANCH > renpy/dl/$PY3_VERSION/branch.txt
+    echo $BRANCH > renpy/dl/$PY2_VERSION/branch.txt
+    popd
+fi
 
-popd
+popd # $BASE/renpy
 
 # Make symlinks.
 pushd $BASE/renpy/dl
 
-weblink $PY2_VERSION current-7-$BRANCH
 weblink $PY3_VERSION current-8-$BRANCH
+weblink $PY2_VERSION current-7-$BRANCH
 
 if [ $BRANCH = main -o $BRANCH = master ]; then
-    weblink $PY2_VERSION current-7
-    weblink $PY3_VERSION current-8
     weblink $PY3_VERSION current
+    weblink $PY3_VERSION current-8
+    weblink $PY2_VERSION current-7
 fi
 
 popd
